refactor(spa-editor): collapse StepCard sub-files (Pass H)
Consolidate the StepCard folder from 24 files toward the audit's
target shape:

Format helpers (4 -> 1):
- Merge `format-cadence-pace.ts`, `format-power-hr.ts`,
  `format-helpers.ts`, and `format-target-helpers.ts` into
  `format-target.ts`. Single `asRecord` predicate replaces four
  `typeof value !== "object" || value === null` duplications;
  `isValidNumber` moves inline as a local const.
- Merge `format-cadence-pace.test.ts` + `format-power-hr.test.ts`
  into `format-target.test.ts` with a shared `baseShapeCases`
  array driving the 6 common rejection cases across all four
  formatters (per audit finding #9).

Component (3 -> 2):
- Inline the rendered JSX of `render-step-card-content.tsx` and
  `get-step-label.ts` directly into `StepCard.tsx`, eliminating
  the anti-pattern of a hook (`useStepCardData`) returning JSX.
- Move the pure derivations (`extractHtmlProps`,
  `deriveStepCardData`) into a sibling `step-card-data.ts` so
  `StepCard.tsx` stays under the 80-line file limit.
