--- packages/workout-spa-editor/src/components/molecules/StepCard/format-target-helpers.ts
@@ -1,5 +1,0 @@
-/**
- * Format target helpers - Re-exports for modular structure
- */
-export { formatCadenceTarget, formatPaceTarget } from "./format-cadence-pace";
-export { formatHeartRateTarget, formatPowerTarget } from "./format-power-hr";
