--- packages/workout-spa-editor/src/components/molecules/StepCard/use-step-card-data.ts
@@ -1,46 +1,0 @@
-import { getStepLabel } from "./get-step-label";
-import { renderStepCardContent } from "./render-step-card-content";
-import type { StepCardProps } from "./StepCard.types";
-import { getStepCardClasses } from "./use-step-card-classes";
-
-/** Compute derived data and content for StepCard rendering. */
-export function useStepCardData(props: StepCardProps) {
-  const {
-    step,
-    visualIndex,
-    isSelected = false,
-    isMultiSelected = false,
-    onDelete,
-    onDuplicate,
-    onCopy,
-    isDragging = false,
-    dragHandleProps,
-    className = "",
-  } = props;
-
-  const selected = isSelected || isMultiSelected;
-  const displayIndex = visualIndex ?? step.stepIndex;
-  const hasActions = Boolean(onDelete || onDuplicate || onCopy);
-
-  return {
-    selected,
-    displayIndex,
-    label: getStepLabel(displayIndex, step.name, step),
-    classes: getStepCardClasses(
-      selected,
-      hasActions,
-      Boolean(dragHandleProps),
-      className
-    ),
-    content: renderStepCardContent({
-      step,
-      displayIndex,
-      intensity: step.intensity || "other",
-      dragHandleProps,
-      isDragging,
-      onDelete,
-      onDuplicate,
-      onCopy,
-    }),
-  };
-}
